+++ entities/src/main/java/com/wowlet/entities/local/SendTransactionModel.kt
@@ -2,7 +2,7 @@ package com.wowlet.entities.local
 
 data class SendTransactionModel(
     val toPublickKey: String,
-    val lamports: Int=0,
+    val lamports: Long=0,
     var fromPublicKey:String="",
     var secretKey:String=""
